Set app debug to False
App debug should be set to false when deploying application to a public server.  Ok to set it to True if running on localhost but it should not be True for a production app.  Will set it to False in the repo so it will be the default if the repo is cloned or if the code is used by someone else.